+++ include/mcrpc/mcrpc.h
@@ -16,6 +16,8 @@
 #include <mcrpc/Parser.h>
 #include <mcrpc/InboundMessage.h>
 #include <mcrpc/OutboundBuilder.h>
+#include <mcrpc/CallResult.h>
+#include <mcrpc/UptimeFormat.h>
 #include <mcrpc/StatusBuilder.h>
 #include <mcrpc/DiscoverBuilder.h>
 #include <mcrpc/CommandRegistry.h>